Cleanup domain models from request serialization
