+++ tests/unittests/test_request_serializer.py
@@ -1,4 +1,6 @@
-from typing import Any, Mapping
+import json
+from datetime import datetime
+from typing import Any, Mapping, Optional
 
 import pytest
 from pydantic import SecretStr
@@ -11,7 +13,13 @@ from blacksmith import (
     QueryStringField,
     Request,
 )
-from blacksmith.service.request_serializer import serialize_request
+from blacksmith.service.request_serializer import (
+    QUERY,
+    JSONEncoder,
+    get_location,
+    serialize_part,
+    serialize_request,
+)
 
 
 class DummyGetRequest(Request):
@@ -22,6 +30,91 @@ class DummyGetRequest(Request):
 
 class DummyPostRequest(DummyGetRequest):
     foo: str = PostBodyField()
+
+
+def test_json_encoder() -> None:
+    assert (
+        json.dumps({"date": datetime(2020, 10, 5)}, cls=JSONEncoder)
+        == '{"date": "2020-10-05T00:00:00"}'
+    )
+
+    with pytest.raises(TypeError) as ctx:
+        json.dumps({"oops": object()}, cls=JSONEncoder)
+    assert str(ctx.value) == "Object of type object is not JSON serializable"
+
+
+def test_get_location_from_pydantic_v2() -> None:
+    class DummyFieldInfo:
+        json_schema_extra = {"location": QUERY}
+
+    assert get_location(DummyFieldInfo()) == QUERY
+
+
+def test_get_location_from_pydantic_v1() -> None:
+    class DummyFieldInfo:
+        class field_info:
+            extra = {"location": QUERY}
+
+    assert get_location(DummyFieldInfo()) == QUERY
+
+
+def test_get_location_raises_value_error() -> None:
+    class Dummy:
+        def __str__(self):
+            return "dummy"
+
+    with pytest.raises(ValueError) as ctx:
+        get_location(Dummy())
+    assert str(ctx.value) == "dummy is not a FieldInfo"
+
+
+def test_serialize_part() -> None:
+    class Dummy(Request):
+        x_message_id: int = HeaderField(default=123, alias="X-Message-Id")
+        name: str = PostBodyField()
+        age: int = PostBodyField(default=10)
+        city: Optional[str] = PostBodyField(None)
+        state: Optional[str] = PostBodyField(None)
+        country: str = PostBodyField()
+
+    dummy = Dummy(name="Jane", country="FR", city="Saint Palais s/mer", state=None)
+    obj = serialize_part(
+        dummy,
+        {
+            "name": ...,
+            "age": ...,
+            "city": ...,
+            "state": ...,
+            "country": ...,
+        },
+    )
+    assert obj == {
+        "name": "Jane",
+        "age": 10,
+        "city": "Saint Palais s/mer",
+        "state": None,
+        "country": "FR",
+    }
+
+
+def test_serialize_part_default_with_none() -> None:
+    class Dummy(Request):
+        name: str = PostBodyField()
+        age: Optional[int] = PostBodyField(default=10)
+
+    dummy = Dummy(name="Jane", age=None)
+    obj = serialize_part(
+        dummy,
+        {
+            "name": ...,
+            "age": ...,
+            "created_at": ...,
+        },
+    )
+    assert obj == {
+        "name": "Jane",
+        "age": None,
+    }
 
 
 @pytest.mark.parametrize(
